requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autopep8==2.3.2
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Django==5.1.5
django-cors-headers==4.7.0
django-filter==24.3
django-redis==5.4.0
django-rest-framework-roles==0.6
django-rest-knox==5.0.2
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-serializer-extensions==2.0.1
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
gprof2dot==2024.6.6
hashids==1.3.1
idna==3.10
inflection==0.5.1
isort==6.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pika==1.3.2
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
